Langevin/requirement.txt
python                    3.8.3                hcff3b4d_0  
numpy                     1.19.2           py38h54aff64_0  
pytorch                   1.6.0               py3.8_cpu_0  [cpuonly]  pytorch
scipy                     1.5.2            py38h0b6359f_0  
h5py                      2.10.0           py38hd6299e0_1  
scikit-learn              0.23.2           py38h0573a6f_0  
pandas                    1.2.4            py38ha9443f7_0
matplotlib                3.1.3                    py38_0
